Refactor document generation for e-mail generation
